extract upload resources modal into its own component
